+++ test/loiste/core_test.clj
@@ -95,7 +95,7 @@
     file
 
     (let [read-workbook (l/workbook file)]
-      (is (= [{:Stuff "This is a long text that should wrap" :Date #inst "2016-03-09T14:05:00"}
+      (is (= [{:Stuff "This is a long text that should wrap" :Date #inst "2016-03-09T14:05:00" :Money nil}
               {:Stuff "Bar" :Date #inst "2016-03-09T14:05:00" :Money 15.0}
               {:Stuff "Qux" :Date #inst "2173-10-13T21:00:00.000-00:00" :Money 100.0}
               {:Stuff "Integer" :Date #inst "1900-01-02T22:20:11.000-00:00" :Money 10000.0}